# docs/swagger.yaml
basePath: /
definitions:
...
  model.CreateHarborRequest:
    properties:
      anchorage_depth:
        minimum: 0
        type: number
      berth_count:
        minimum: 0
        type: integer
      cargo_handling_capacity:
        minimum: 0
        type: number
      channel_depth:
        minimum: 0
        type: number
      city:
        maxLength: 100
        type: string
      contact_email:
        maxLength: 255
        type: string
      contact_person:
        maxLength: 255
        type: string
      contact_phone:
        maxLength: 20
        type: string
      country:
        maxLength: 100
        type: string
      harbor_code:
        maxLength: 20
        type: string
      harbor_name:
        maxLength: 255
        type: string
      has_breakbulk:
        type: boolean
      has_bulk:
        type: boolean
      has_cargo:
        type: boolean
      has_container:
        type: boolean
      has_customs:
        type: boolean
      has_fuel:
        type: boolean
      has_immigration:
        type: boolean
      has_liquid:
        type: boolean
      has_medical:
        type: boolean
      has_passenger:
        type: boolean
      has_pilotage:
        type: boolean
      has_quarantine:
        type: boolean
      has_repair:
        type: boolean
      has_roro:
        type: boolean
      has_security:
        type: boolean
      has_supplies:
        type: boolean
      has_tug_service:
        type: boolean
      has_waste:
        type: boolean
      has_water:
        type: boolean
      latitude:
        maximum: 90
        minimum: -90
        type: number
      longitude:
        maximum: 180
        minimum: -180
        type: number
      max_ship_beam:
        minimum: 0
        type: number
      max_ship_draft:
        minimum: 0
        type: number
      max_ship_length:
        minimum: 0
        type: number
      notes:
        maxLength: 1000
        type: string
      operating_hours:
        maxLength: 255
        type: string
      province:
        maxLength: 100
        type: string
      storage_capacity:
        minimum: 0
        type: number
      time_zone:
        maxLength: 50
        type: string
      un_locode:
        maxLength: 10
        type: string
      website:
        maxLength: 500
        type: string
    required:
    - city
    - country
    - harbor_code
    - harbor_name
    - province
    type: object
...
  model.PageMetadata:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
...
  model.SwaggerPageResponse:
    properties:
      data:
        items: {}
        type: array
      errors:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/model.PageMetadata'
      success:
        type: boolean
    type: object
  model.SwaggerWebResponse:
    properties:
      data: {}
      errors:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/model.PageMetadata'
      success:
        type: boolean
    type: object
  model.UpdateHarborRequest:
    properties:
      anchorage_depth:
        minimum: 0
        type: number
      berth_count:
        minimum: 0
        type: integer
      cargo_handling_capacity:
        minimum: 0
        type: number
      channel_depth:
        minimum: 0
        type: number
      city:
        maxLength: 100
        type: string
      contact_email:
        maxLength: 255
        type: string
      contact_person:
        maxLength: 255
        type: string
      contact_phone:
        maxLength: 20
        type: string
      country:
        maxLength: 100
        type: string
      harbor_code:
        maxLength: 20
        type: string
      harbor_name:
        maxLength: 255
        type: string
      has_breakbulk:
        type: boolean
      has_bulk:
        type: boolean
      has_cargo:
        type: boolean
      has_container:
        type: boolean
      has_customs:
        type: boolean
      has_fuel:
        type: boolean
      has_immigration:
        type: boolean
      has_liquid:
        type: boolean
      has_medical:
        type: boolean
      has_passenger:
        type: boolean
      has_pilotage:
        type: boolean
      has_quarantine:
        type: boolean
      has_repair:
        type: boolean
      has_roro:
        type: boolean
      has_security:
        type: boolean
      has_supplies:
        type: boolean
      has_tug_service:
        type: boolean
      has_waste:
        type: boolean
      has_water:
        type: boolean
      is_active:
        type: boolean
      latitude:
        maximum: 90
        minimum: -90
        type: number
      longitude:
        maximum: 180
        minimum: -180
        type: number
      max_ship_beam:
        minimum: 0
        type: number
      max_ship_draft:
        minimum: 0
        type: number
      max_ship_length:
        minimum: 0
        type: number
      notes:
        maxLength: 1000
        type: string
      operating_hours:
        maxLength: 255
        type: string
      province:
        maxLength: 100
        type: string
      storage_capacity:
        minimum: 0
        type: number
      time_zone:
        maxLength: 50
        type: string
      un_locode:
        maxLength: 10
        type: string
      website:
        maxLength: 500
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Golang Clean Architecture.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MKP Boarding Test API
  version: "1.0"
paths:
  /api/harbors:
    get:
      consumes:
      - application/json
      description: Get list of harbors with optional filtering
      parameters:
      - description: Filter by harbor name
        in: query
        name: name
        type: string
      - description: Filter by country
        in: query
        name: country
        type: string
      - description: Filter by province
        in: query
        name: province
        type: string
      - description: Filter by city
        in: query
        name: city
        type: string
      - default: true
        description: Filter by active status
        in: query
        name: is_active
        type: boolean
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of harbors
          schema:
            $ref: '#/definitions/model.SwaggerPageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: List harbors
      tags:
      - Harbors
    post:
      consumes:
      - application/json
      description: Create a new harbor with detailed information
      parameters:
      - description: Create harbor request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateHarborRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Harbor created successfully
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Create a new harbor
      tags:
      - Harbors
  /api/harbors/{harborId}:
    delete:
      consumes:
      - application/json
      description: Delete harbor by harbor ID
      parameters:
      - description: Harbor ID
        in: path
        name: harborId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Harbor deleted successfully
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "404":
          description: Harbor not found
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Delete harbor
      tags:
      - Harbors
    get:
      consumes:
      - application/json
      description: Get harbor details by harbor ID
      parameters:
      - description: Harbor ID
        in: path
        name: harborId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Harbor details
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "404":
          description: Harbor not found
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Get harbor by ID
      tags:
      - Harbors
    put:
      consumes:
      - application/json
      description: Update harbor information by harbor ID
      parameters:
      - description: Harbor ID
        in: path
        name: harborId
        required: true
        type: string
      - description: Update harbor request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateHarborRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Harbor updated successfully
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "404":
          description: Harbor not found
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Update harbor
      tags:
      - Harbors
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"